requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1 ; python_version >= '3.8'
certifi==2025.1.31 ; python_version >= '3.6'
charset-normalizer==3.4.1 ; python_version >= '3.7'
django==5.2
django-ckeditor==6.7.2
django-js-asset==3.1.2 ; python_version >= '3.10'
idna==3.10 ; python_version >= '3.6'
mercadopago==2.3.0
pillow==11.2.1
requests==2.32.3 ; python_version >= '3.8'
sqlparse==0.5.3 ; python_version >= '3.8'
urllib3==2.4.0 ; python_version >= '3.9'
